requirements.txt
Authlib==0.15.4
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-dotenv==0.19.0
Werkzeug==2.0.1
